[WIP] Continue extracting mutators
Pivot to preserve services. Currently, the separation is Mutation
classes as interface (controller?), Mutator class as a request
authorization and validation layer (might need a better name), and
Service class as _executors_. Reform is being ditched because it
expects that there would be an input _object_, which the Mutation layer
would have to resolve - but also involves authorization.

This clarifies responsibility at each layer - hopefully I won't change
my mind in a few days.